requirements.txt
openai>=1.12.0
google-generativeai>=0.3.0
google-genai>=0.2.0
requests>=2.31.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
opencv-python>=4.8.0
pillow>=10.0.0
python-dotenv>=1.0.0
boto3>=1.34.0

